requirements.txt
fabric
invoke
PyYAML
requests
urllib3
paramiko